+++ modules/uhsg_search/js/uhsg-search.js
@@ -1,16 +1,14 @@
 (function ($) {
-  'use strict';
-  Drupal.behaviors.uhsg_search_reset = {
-    attach: function(context, settings) {
-
-      // Exposed filter reset redirects user to 404 page on AJAX view when
-      // placed as a block (https://www.drupal.org/node/2820347). Until core is
-      // fixed, using location.reload() to reload the current page to reset the
-      // form.
-      $('.view-search .views-exposed-form input.button--reset').click(function (e) {
-        e.preventDefault();
-    	  location.reload();
-      })
+  // Submit search when clicking suggestions.
+  var oldSelect = Drupal.autocomplete.options.select;
+  Drupal.autocomplete.options.select = function(event, ui) {
+    oldSelect.call(this, event, ui);
+    if ($(this).hasClass('ui-autocomplete-input')) {
+      $(this).siblings('.form-actions').children('#edit-submit-search').trigger('click');
     }
+  };
+  // Autocomplete requires the user to click twice in iOS, fix that.
+  Drupal.autocomplete.options.open = function (event, ui) {
+    $('.ui-autocomplete').off('menufocus hover mouseover');
   };
 }(jQuery));