+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 #Changelog
 
 ## 1.0-dev
-
+* Fix issues with search autocomplete on mobile devices (HUB-136)
+* Minor security improvements to my searches functionality (HUB-152)
+* Remove theme sorting when using mobile devices (HUB-89)
 * Do not prevent login when getting study rights from oprek fails (HUB-153)